Year 3 Semester 2/Django/Lab 2/requirements.txt
mysql-connector-python==8.0.32
protobuf==3.20.3
psycopg2==2.9.5
tk==0.1.0
